requirements.txt
streamlit
google-generativeai
python-dotenv
sqlite3
sqlparse